multi-agent-orchestration-demo/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
    ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY")
    PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
    PINECONE_INDEX = os.getenv("PINECONE_INDEX", "agent-context")
    REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
    DEFAULT_MODEL = os.getenv("DEFAULT_MODEL", "gpt-4o")
    CONFIDENCE_THRESHOLD = float(os.getenv("CONFIDENCE_THRESHOLD", "0.7"))
    SCHEDULE_INTERVAL_MINUTES = int(os.getenv("SCHEDULE_INTERVAL_MINUTES", "60"))
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
