multi-agent-orchestration-demo/requirements.txt
openclaw>=0.3.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
anthropic>=0.28.0
pinecone-client>=3.0.0
redis>=5.0.0
apscheduler>=3.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
